requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
python-multipart>=0.0.6
httpx>=0.24.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
aiohttp>=3.8.5
aiohttp-retry>=2.8.3
asyncio>=3.4.3
aiofiles>=23.1.0
neo4j>=4.4.0
redis>=4.0.0
aioredis>=2.0.0
pymilvus>=2.3.1
sentence-transformers>=2.2.2
numpy>=1.24.0
scikit-learn>=1.3.2
networkx>=3.1
prometheus-client>=0.19.0
structlog>=21.1.0
fastmcp>=0.3.4
tenacity>=8.0.1
backoff>=2.2.1
cachetools>=5.3.1
python-jose[cryptography]>=3.3.0
passlib>=1.7.4
bcrypt>=3.2.0
pytest>=6.2.5
requests>=2.26.0
docker>=5.0.0
python-dateutil>=2.8.2
typing-extensions>=4.7.1